gfgf

+++ assignment-mpi-numint-collective/num_int/mpi_num_int.cpp
@@ -50,7 +50,7 @@ int main (int argc, char* argv[]) {
   int n = atoi(argv[4]);
   int intensity = atoi(argv[5]);
 
-  double result = 0; // init result
+  float result = 0; // init result
   float (*ptr)(float, int) = getFunction(func); // get function
   float co =  (b - a) / float (n); // calculate coefficient
 
@@ -63,7 +63,7 @@ int main (int argc, char* argv[]) {
 
   int loop_start = rank * (n / size);
   int loop_end = (rank + 1) * (n / size);
-  double rank_val = 0;
+  float rank_val = 0;
 
   if (loop_start + 1 == rank) loop_end = n;
 
@@ -71,7 +71,7 @@ int main (int argc, char* argv[]) {
     rank_val += (*ptr)(a + ((i + .5) * co), intensity);
   }
 
-  MPI_Reduce(&rank_val, &result, size, MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
+  MPI_Reduce(&rank_val, &result, size, MPI_FLOAT, MPI_SUM, 0, MPI_COMM_WORLD);
 
   if (rank == 0){
     result *= co;
